% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraction_from_rasters.R
\name{Extraction_from_rasters}
\alias{Extraction_from_rasters}
\title{Extracting information from rasters file. Example using observed precipitation and temperature for Klementinum station
And simulating data}
\usage{
Extraction_from_rasters(
  path,
  name_of_file_pr,
  name_of_file_tem,
  lon,
  lat,
  name_of_hist_pr,
  name_of_hist_tem,
  n_pr_sim,
  n_tas_sim,
  sim_date_range,
  obs_date_range
)
}
\arguments{
\item{path}{(string) path to your working directory with input files}

\item{name_of_file_pr}{(string) name of your file with daily precipitation based on station}

\item{name_of_file_tem}{(string) name of your file with daily temperature based on station}

\item{lon}{(string) longitude of station in format hhh:mm:ss}

\item{lat}{(string) latitude of station in format hhh:mm:ss}

\item{name_of_hist_pr}{(string) name of your historical data by default used: EU_day_pr}

\item{name_of_hist_tem}{(string) name of your historical data by default used: EU_day_tas}

\item{n_pr_sim}{(string) name of presimulated data for precipitation}

\item{n_tas_sim}{(string) name of presimulated data for temperature}

\item{sim_date_range}{(list) range of period for future simulation or scenario period can be start from nowadays till 2099 year.}

\item{obs_date_range}{(list) range of period for observation data}
}
\value{
list with prepared data include: simulation data, observed data
}
\description{
Extracting information from rasters file. Example using observed precipitation and temperature for Klementinum station
And simulating data
}
\examples{
name_of_file_pr = system.file("extdata", "RR_STAID000027.txt", package = "CCSusingADCmethod")
name_of_file_tem = system.file("extdata", "TG_STAID000027.txt", package = "CCSusingADCmethod")
lon = "+014:24:59"
lat = "+50:05:11"
name_of_hist_pr = system.file("extdata", "EUR_pr_day_CanESM2_historical_rcp85_r25i1p1_19500101-21001231.nc", package = "CCSusingADCmethod")
name_of_hist_tem = system.file("extdata", "EUR_tas_day_CanESM2_historical_rcp85_r50i1p1_19500101-21001231.nc", package = "CCSusingADCmethod")
n_pr_sim = system.file("extdata", "pr_sim.rds", package = "CCSusingADCmethod")
n_tas_sim = system.file("extdata", "tas_sim.rds", package = "CCSusingADCmethod")
sim_date_range = c(1951:2099)
obs_date_range = c(1805:2004)
Example_output_extr = Extraction_from_rasters(, name_of_file_pr, name_of_file_tem, lon, lat, name_of_hist_pr, name_of_hist_tem, n_pr_sim, n_tas_sim, sim_date_range, obs_date_range)
}
