% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_to_CCS.R
\name{obs_to_CCS}
\alias{obs_to_CCS}
\title{Obs_to_CCS}
\usage{
obs_to_CCS(obs, stat, obs_date_range)
}
\arguments{
\item{obs}{prepared observed data}

\item{stat}{transformation parameters}

\item{obs_date_range}{(list) range of period for observation data}
}
\value{
Climate Change Scenario in form data table
}
\description{
Transformation of the observed series into the climate change scenario
}
\examples{
obs = Example_output_extr[["obs"]]
stat = Example_output_trans_param
obs_date_range = c(1805:2004)
ccs = obs_to_CCS(obs, stat, obs_date_range)

}
