% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R, R/Plotting_pr_trans.R
\name{pr_trans_ploting}
\alias{pr_trans_ploting}
\title{pr_trans_ploting}
\usage{
pr_trans_ploting(ccs, t_period)

pr_trans_ploting(ccs, t_period)
}
\arguments{
\item{ccs}{data table with Climate Change Scenario}

\item{t_period}{time period for plotting}
}
\value{
Graph with transformed precipitation(time_period)

Graph with transformed precipitation(time_period)
}
\description{
Results plotting functions

Results plotting functions
}
