% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_sd_r.R
\name{sd_r_ploting}
\alias{sd_r_ploting}
\title{sd_r_ploting}
\usage{
sd_r_ploting(ccs, t_period)
}
\arguments{
\item{ccs}{data table with Climate Change Scenario}

\item{t_period}{time period for plotting}
}
\value{
Graph with transformed precipitation(time_period)
}
\description{
Results plotting functions
}
