% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_tas_trans.R
\name{tas_trans_ploting}
\alias{tas_trans_ploting}
\title{tas_trans_ploting}
\usage{
tas_trans_ploting(ccs, t_period)
}
\arguments{
\item{ccs}{data table with Climate Change Scenario}

\item{t_period}{time period for plotting}
}
\value{
Graph with transformed precipitation(time_period)
}
\description{
Results plotting functions
}
