% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation_param.R
\name{Tranformation_param}
\alias{Tranformation_param}
\title{Getting the coefficients a and b were derived from the 60% quantile (P60)
and the 90% quantile (P90) of the 5-day precipitation sums
and the (future) changes therein.}
\usage{
Tranformation_param(sim, obs, obs_date_range, sim_date_range)
}
\arguments{
\item{sim}{Data Table with prepared simulation data}

\item{obs}{Data Table with prepared observed data}

\item{obs_date_range}{(list) range of period for observation data}

\item{sim_date_range}{(list) range of period for future simulation or scenario period can be start from nowadays till 2099 year.}
}
\value{
stats data, transormation parameters
}
\description{
Getting the coefficients a and b were derived from the 60% quantile (P60)
and the 90% quantile (P90) of the 5-day precipitation sums
and the (future) changes therein.
}
\examples{
sim = Example_output_extr[["sim"]]
obs = Example_output_extr[["obs"]]
obs_date_range = c(1805:2004)
sim_date_range = c(1951:2099)
Example_output_trans_param = Tranformation_param(sim, obs, obs_date_range, sim_date_range)

}
